+++ weblate/addons/cleanup.py
@@ -25,31 +25,15 @@ import os
 from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
 from translate.storage.lisa import LISAfile
-from translate.storage.resx import RESXFile
 
 from weblate.addons.base import UpdateBaseAddon
 from weblate.formats.txt import AppStoreParser
 
 
-class CleanupAddon(UpdateBaseAddon):
-    name = 'weblate.cleanup.generic'
-    verbose = _('Cleanup translation files')
-    description = _(
-        'Update all translation files to match the monolingual base file. '
-        'For most file formats, this means removing stale translation keys '
-        'no longer present in the base file.'
-    )
-    icon = 'eraser'
-
+class BaseCleanupAddon(UpdateBaseAddon):
     @cached_property
     def template_store(self):
         return self.instance.component.template_store.store
-
-    @classmethod
-    def can_install(cls, component, user):
-        if not component.has_template():
-            return False
-        return super(CleanupAddon, cls).can_install(component, user)
 
     @staticmethod
     def build_index(storage):
@@ -59,6 +43,30 @@ class CleanupAddon(UpdateBaseAddon):
             index[unit.getid()] = unit
 
         return index
+
+    @staticmethod
+    def iterate_translations(component):
+        for translation in component.translation_set.iterator():
+            if translation.is_template:
+                continue
+            yield translation
+
+    @classmethod
+    def can_install(cls, component, user):
+        if not component.has_template():
+            return False
+        return super(BaseCleanupAddon, cls).can_install(component, user)
+
+
+class CleanupAddon(BaseCleanupAddon):
+    name = 'weblate.cleanup.generic'
+    verbose = _('Cleanup translation files')
+    description = _(
+        'Update all translation files to match the monolingual base file. '
+        'For most file formats, this means removing stale translation keys '
+        'no longer present in the base file.'
+    )
+    icon = 'eraser'
 
     def update_appstore(self, index, translation, storage):
         """Filter obsolete units in storage.
@@ -100,55 +108,6 @@ class CleanupAddon(UpdateBaseAddon):
         if changed:
             storage.save()
 
-    def update_resx(self, index, translation, storage, changes):
-        """Filter obsolete units in RESX storage.
-
-        This removes the corresponding XML element and
-        also adds newly added units and changed ones.
-        """
-        sindex = self.build_index(storage)
-        changed = False
-
-        # Add missing units
-        for unit in self.template_store.units:
-            if unit.getid() not in sindex:
-                storage.addunit(unit, True)
-                changed = True
-
-        # Remove extra units and apply target changes
-        for unit in storage.units:
-            unitid = unit.getid()
-            if unitid not in index:
-                storage.body.remove(unit.xmlelement)
-                changed = True
-            if unitid in changes:
-                unit.target = index[unitid].target
-                changed = True
-
-        if changed:
-            storage.save()
-
-    @staticmethod
-    def iterate_translations(component):
-        for translation in component.translation_set.iterator():
-            if translation.is_template:
-                continue
-            yield translation
-
-    @staticmethod
-    def find_changes(index, storage):
-        """Find changed units in storage"""
-        result = set()
-
-        for unit in storage.units:
-            unitid = unit.getid()
-            if unitid not in index:
-                continue
-            if unit.target != index[unitid].target:
-                result.add(unitid)
-
-        return result
-
     def update_translations(self, component, previous_head):
         index = self.build_index(self.template_store)
 
@@ -156,22 +115,6 @@ class CleanupAddon(UpdateBaseAddon):
             for translation in self.iterate_translations(component):
                 self.update_appstore(
                     index, translation, translation.store.store
-                )
-        elif isinstance(self.template_store, RESXFile):
-            if previous_head:
-                content = component.repository.get_file(
-                    component.template, previous_head
-                )
-                changes = self.find_changes(
-                    index,
-                    RESXFile.parsestring(content)
-                )
-            else:
-                # No previous revision, probably first commit
-                changes = set()
-            for translation in self.iterate_translations(component):
-                self.update_resx(
-                    index, translation, translation.store.store, changes
                 )
         elif isinstance(self.template_store, LISAfile):
             for translation in self.iterate_translations(component):