+++ weblate/settings_example.py
@@ -763,6 +763,7 @@ CRISPY_TEMPLATE_PACK = 'bootstrap3'
 #     'weblate.addons.git.GitSquashAddon',
 #     'weblate.addons.removal.RemoveComments',
 #     'weblate.addons.removal.RemoveSuggestions',
+#     'weblate.addons.resx.ResxUpdateAddon',
 #     'weblate.addons.autotranslate.AutoTranslateAddon',
 # )